+++ banyand/liaison/grpc/server.go
@@ -22,6 +22,7 @@ import (
 	"context"
 	"net"
 	"runtime/debug"
+	"strconv"
 	"time"
 
 	"github.com/grpc-ecosystem/go-grpc-middleware/v2/interceptors/recovery"
@@ -56,33 +57,41 @@ var (
 	errAccessLogRootPath = errors.New("access log root path is required")
 )
 
+// Server defines the gRPC server.
+type Server interface {
+	run.Unit
+	GetPort() *uint32
+}
+
 type server struct {
-	pipeline queue.Queue
+	pipeline queue.Client
 	creds    credentials.TransportCredentials
-	*indexRuleRegistryServer
-	measureSVC *measureService
-	log        *logger.Logger
-	ser        *grpclib.Server
+	*streamRegistryServer
+	log *logger.Logger
+	*indexRuleBindingRegistryServer
+	ser *grpclib.Server
 	*propertyServer
 	*topNAggregationRegistryServer
 	*groupRegistryServer
-	stopCh    chan struct{}
-	streamSVC *streamService
+	stopCh chan struct{}
+	*indexRuleRegistryServer
 	*measureRegistryServer
-	*streamRegistryServer
-	*indexRuleBindingRegistryServer
-	addr                     string
+	streamSVC                *streamService
+	measureSVC               *measureService
+	host                     string
 	keyFile                  string
 	certFile                 string
 	accessLogRootPath        string
+	addr                     string
 	accessLogRecorders       []accessLogRecorder
 	maxRecvMsgSize           run.Bytes
-	tls                      bool
+	port                     uint32
 	enableIngestionAccessLog bool
+	tls                      bool
 }
 
 // NewServer returns a new gRPC server.
-func NewServer(_ context.Context, pipeline queue.Queue, schemaRegistry metadata.Repo) run.Unit {
+func NewServer(_ context.Context, pipeline queue.Client, schemaRegistry metadata.Repo) Server {
 	streamSVC := &streamService{
 		discoveryService: newDiscoveryService(pipeline, schema.KindStream, schemaRegistry),
 	}
@@ -119,7 +128,7 @@ func NewServer(_ context.Context, pipeline queue.Queue, schemaRegistry metadata.
 	return s
 }
 
-func (s *server) PreRun() error {
+func (s *server) PreRun(ctx context.Context) error {
 	s.log = logger.GetLogger("liaison-grpc")
 	s.streamSVC.setLogger(s.log)
 	s.measureSVC.setLogger(s.log)
@@ -129,7 +138,7 @@ func (s *server) PreRun() error {
 	}
 	for _, c := range components {
 		c.SetLogger(s.log)
-		if err := c.initialize(); err != nil {
+		if err := c.initialize(ctx); err != nil {
 			return err
 		}
 	}
@@ -148,6 +157,14 @@ func (s *server) Name() string {
 	return "grpc"
 }
 
+func (s *server) Role() databasev1.Role {
+	return databasev1.Role_ROLE_LIAISON
+}
+
+func (s *server) GetPort() *uint32 {
+	return &s.port
+}
+
 func (s *server) FlagSet() *run.FlagSet {
 	fs := run.NewFlagSet("grpc")
 	s.maxRecvMsgSize = defaultRecvSize
@@ -155,14 +172,16 @@ func (s *server) FlagSet() *run.FlagSet {
 	fs.BoolVar(&s.tls, "tls", false, "connection uses TLS if true, else plain TCP")
 	fs.StringVar(&s.certFile, "cert-file", "", "the TLS cert file")
 	fs.StringVar(&s.keyFile, "key-file", "", "the TLS key file")
-	fs.StringVar(&s.addr, "addr", ":17912", "the address of banyand listens")
+	fs.StringVar(&s.host, "grpc-host", "", "the host of banyand listens")
+	fs.Uint32Var(&s.port, "grpc-port", 17912, "the port of banyand listens")
 	fs.BoolVar(&s.enableIngestionAccessLog, "enable-ingestion-access-log", false, "enable ingestion access log")
 	fs.StringVar(&s.accessLogRootPath, "access-log-root-path", "", "access log root path")
 	return fs
 }
 
 func (s *server) Validate() error {
-	if s.addr == "" {
+	s.addr = net.JoinHostPort(s.host, strconv.FormatUint(uint64(s.port), 10))
+	if s.addr == ":" {
 		return errNoAddr
 	}
 	if s.enableIngestionAccessLog && s.accessLogRootPath == "" {