+++ banyand/measure/service.go
@@ -20,10 +20,8 @@ package measure
 import (
 	"context"
 	"path"
-	"time"
 
 	"github.com/pkg/errors"
-	"go.uber.org/multierr"
 
 	"github.com/apache/skywalking-banyandb/api/data"
 	commonv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/common/v1"
@@ -59,7 +57,8 @@ type service struct {
 	schemaRepo             schemaRepo
 	writeListener          bus.MessageListener
 	metadata               metadata.Repo
-	pipeline               queue.Queue
+	pipeline               queue.Server
+	localPipeline          queue.Queue
 	l                      *logger.Logger
 	root                   string
 	dbOpts                 tsdb.DatabaseOpts
@@ -103,79 +102,28 @@ func (s *service) Name() string {
 	return "measure"
 }
 
-func (s *service) PreRun() error {
+func (s *service) Role() databasev1.Role {
+	return databasev1.Role_ROLE_DATA
+}
+
+func (s *service) PreRun(_ context.Context) error {
 	s.l = logger.GetLogger(s.Name())
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	groups, err := s.metadata.GroupRegistry().ListGroup(ctx)
-	cancel()
-	if err != nil {
-		return err
-	}
 	path := path.Join(s.root, s.Name())
 	observability.UpdatePath(path)
+	s.localPipeline = queue.Local()
 	s.schemaRepo = newSchemaRepo(path, s.metadata, s.dbOpts,
-		s.l, s.pipeline, int64(s.BlockEncoderBufferSize), int64(s.BlockBufferSize))
-	for _, g := range groups {
-		if g.Catalog != commonv1.Catalog_CATALOG_MEASURE {
-			continue
-		}
-		gp, innerErr := s.schemaRepo.StoreGroup(g.Metadata)
-		if innerErr != nil {
-			return innerErr
-		}
-		ctx, cancel = context.WithTimeout(context.Background(), 5*time.Second)
-		allMeasureSchemas, innerErr := s.metadata.MeasureRegistry().
-			ListMeasure(ctx, schema.ListOpt{Group: gp.GetSchema().GetMetadata().GetName()})
-		cancel()
-		if innerErr != nil {
-			return innerErr
-		}
-		for _, measureSchema := range allMeasureSchemas {
-			// sanity check before calling StoreResource
-			// since StoreResource may be called inside the event loop
-			if checkErr := s.sanityCheck(gp, measureSchema); checkErr != nil {
-				return checkErr
-			}
-			if _, innerErr := gp.StoreResource(measureSchema); innerErr != nil {
-				return innerErr
-			}
-		}
-	}
+		s.l, s.localPipeline, int64(s.BlockEncoderBufferSize), int64(s.BlockBufferSize))
 	// run a serial watcher
 	go s.schemaRepo.Watcher()
 	s.metadata.
-		RegisterHandler(schema.KindGroup|schema.KindMeasure|schema.KindIndexRuleBinding|schema.KindIndexRule|schema.KindTopNAggregation,
+		RegisterHandler("measure", schema.KindGroup|schema.KindMeasure|schema.KindIndexRuleBinding|schema.KindIndexRule|schema.KindTopNAggregation,
 			&s.schemaRepo)
-
 	s.writeListener = setUpWriteCallback(s.l, &s.schemaRepo)
-	err = s.pipeline.Subscribe(data.TopicMeasureWrite, s.writeListener)
+	err := s.pipeline.Subscribe(data.TopicMeasureWrite, s.writeListener)
 	if err != nil {
 		return err
 	}
-	return nil
-}
-
-func (s *service) sanityCheck(group resourceSchema.Group, measureSchema *databasev1.Measure) error {
-	var topNAggrs []*databasev1.TopNAggregation
-	ctx, cancel := context.WithTimeout(context.Background(), 10*time.Second)
-	defer cancel()
-	topNAggrs, err := s.metadata.MeasureRegistry().TopNAggregations(ctx, measureSchema.GetMetadata())
-	if err != nil || len(topNAggrs) == 0 {
-		return err
-	}
-
-	for _, topNAggr := range topNAggrs {
-		topNMeasure, innerErr := createOrUpdateTopNMeasure(s.metadata.MeasureRegistry(), topNAggr)
-		err = multierr.Append(err, innerErr)
-		if topNMeasure != nil {
-			_, storeErr := group.StoreResource(topNMeasure)
-			if storeErr != nil {
-				err = multierr.Append(err, storeErr)
-			}
-		}
-	}
-
-	return err
+	return s.localPipeline.Subscribe(data.TopicMeasureWrite, s.writeListener)
 }
 
 func (s *service) Serve() run.StopNotify {
@@ -183,11 +131,12 @@ func (s *service) Serve() run.StopNotify {
 }
 
 func (s *service) GracefulStop() {
+	s.localPipeline.GracefulStop()
 	s.schemaRepo.Close()
 }
 
 // NewService returns a new service.
-func NewService(_ context.Context, metadata metadata.Repo, pipeline queue.Queue) (Service, error) {
+func NewService(_ context.Context, metadata metadata.Repo, pipeline queue.Server) (Service, error) {
 	return &service{
 		metadata: metadata,
 		pipeline: pipeline,