+++ banyand/metadata/client.go
@@ -23,7 +23,9 @@ import (
 	"time"
 
 	"go.uber.org/multierr"
+	"google.golang.org/protobuf/types/known/timestamppb"
 
+	"github.com/apache/skywalking-banyandb/api/common"
 	commonv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/common/v1"
 	databasev1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/database/v1"
 	"github.com/apache/skywalking-banyandb/banyand/metadata/schema"
@@ -62,15 +64,37 @@ func (s *clientService) Validate() error {
 	return nil
 }
 
-func (s *clientService) PreRun() error {
+func (s *clientService) PreRun(ctx context.Context) error {
 	var err error
 	s.schemaRegistry, err = schema.NewEtcdSchemaRegistry(schema.ConfigureServerEndpoints(s.endpoints))
 	if err != nil {
 		return err
 	}
-
+	val := ctx.Value(common.ContextNodeKey)
+	if val == nil {
+		return errors.New("node id is empty")
+	}
+	node := val.(common.Node)
+	val = ctx.Value(common.ContextNodeRolesKey)
+	if val == nil {
+		return errors.New("node roles is empty")
+	}
+	nodeRoles := val.([]databasev1.Role)
+	ctxRegister, cancel := context.WithTimeout(ctx, time.Second*5)
+	defer cancel()
+	if err = s.schemaRegistry.RegisterNode(ctxRegister, &databasev1.Node{
+		Metadata: &commonv1.Metadata{
+			Name: node.NodeID,
+		},
+		GrpcAddress: node.GrpcAddress,
+		HttpAddress: node.HTTPAddress,
+		Roles:       nodeRoles,
+		CreatedAt:   timestamppb.Now(),
+	}); err != nil {
+		return err
+	}
 	s.alc = newAllocator(s.schemaRegistry, logger.GetLogger(s.Name()).Named("allocator"))
-	s.schemaRegistry.RegisterHandler(schema.KindGroup|schema.KindNode, s.alc)
+	s.schemaRegistry.RegisterHandler("shard-allocator", schema.KindGroup|schema.KindNode, s.alc)
 	return nil
 }
 
@@ -84,8 +108,8 @@ func (s *clientService) GracefulStop() {
 	_ = s.schemaRegistry.Close()
 }
 
-func (s *clientService) RegisterHandler(kind schema.Kind, handler schema.EventHandler) {
-	s.schemaRegistry.RegisterHandler(kind, handler)
+func (s *clientService) RegisterHandler(name string, kind schema.Kind, handler schema.EventHandler) {
+	s.schemaRegistry.RegisterHandler(name, kind, handler)
 }
 
 func (s *clientService) StreamRegistry() schema.Stream {
@@ -122,6 +146,10 @@ func (s *clientService) ShardRegistry() schema.Shard {
 
 func (s *clientService) Name() string {
 	return "metadata"
+}
+
+func (s *clientService) Role() databasev1.Role {
+	return databasev1.Role_ROLE_META
 }
 
 func (s *clientService) IndexRules(ctx context.Context, subject *commonv1.Metadata) ([]*databasev1.IndexRule, error) {