+++ banyand/stream/service.go
@@ -20,12 +20,12 @@ package stream
 import (
 	"context"
 	"path"
-	"time"
 
 	"github.com/pkg/errors"
 
 	"github.com/apache/skywalking-banyandb/api/data"
 	commonv1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/common/v1"
+	databasev1 "github.com/apache/skywalking-banyandb/api/proto/banyandb/database/v1"
 	"github.com/apache/skywalking-banyandb/banyand/metadata"
 	"github.com/apache/skywalking-banyandb/banyand/metadata/schema"
 	"github.com/apache/skywalking-banyandb/banyand/observability"
@@ -53,7 +53,7 @@ var _ Service = (*service)(nil)
 type service struct {
 	schemaRepo      schemaRepo
 	metadata        metadata.Repo
-	pipeline        queue.Queue
+	pipeline        queue.Server
 	writeListener   *writeCallback
 	l               *logger.Logger
 	root            string
@@ -93,40 +93,18 @@ func (s *service) Name() string {
 	return "stream"
 }
 
-func (s *service) PreRun() error {
+func (s *service) Role() databasev1.Role {
+	return databasev1.Role_ROLE_DATA
+}
+
+func (s *service) PreRun(_ context.Context) error {
 	s.l = logger.GetLogger(s.Name())
-	ctx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-	groups, err := s.metadata.GroupRegistry().ListGroup(ctx)
-	cancel()
-	if err != nil {
-		return err
-	}
 	path := path.Join(s.root, s.Name())
 	observability.UpdatePath(path)
 	s.schemaRepo = newSchemaRepo(path, s.metadata, int64(s.blockBufferSize), s.dbOpts, s.l)
-	for _, g := range groups {
-		if g.Catalog != commonv1.Catalog_CATALOG_STREAM {
-			continue
-		}
-		gp, err := s.schemaRepo.StoreGroup(g.Metadata)
-		if err != nil {
-			return err
-		}
-		ctx, cancel = context.WithTimeout(context.Background(), 5*time.Second)
-		schemas, err := s.metadata.StreamRegistry().ListStream(ctx, schema.ListOpt{Group: gp.GetSchema().GetMetadata().Name})
-		cancel()
-		if err != nil {
-			return err
-		}
-		for _, sa := range schemas {
-			if _, innerErr := gp.StoreResource(sa); innerErr != nil {
-				return innerErr
-			}
-		}
-	}
 	// run a serial watcher
 	s.schemaRepo.Watcher()
-	s.metadata.RegisterHandler(schema.KindGroup|schema.KindStream|schema.KindIndexRuleBinding|schema.KindIndexRule,
+	s.metadata.RegisterHandler("stream", schema.KindGroup|schema.KindStream|schema.KindIndexRuleBinding|schema.KindIndexRule,
 		&s.schemaRepo)
 
 	s.writeListener = setUpWriteCallback(s.l, &s.schemaRepo)
@@ -147,7 +125,7 @@ func (s *service) GracefulStop() {
 }
 
 // NewService returns a new service.
-func NewService(_ context.Context, metadata metadata.Repo, pipeline queue.Queue) (Service, error) {
+func NewService(_ context.Context, metadata metadata.Repo, pipeline queue.Server) (Service, error) {
 	return &service{
 		metadata: metadata,
 		pipeline: pipeline,