+++ CHANGES.md
@@ -11,6 +11,8 @@ Release Notes.
 - Document the clustering.
 - Support multiple roles for banyand server.
 - Support for recovery buffer using wal.
+- Register the node role to the metadata registry.
+- Implement the remote queue to spreading data to data nodes.
 
 ### Bugs
 
@@ -19,6 +21,7 @@ Release Notes.
 - Fix timer not released
 - BanyanDB ui misses fields when creating a group
 - Fix data duplicate writing
+- Syncing metadata change events from etcd instead of a local channel.
 
 ### Chores